Split the surface ingress service along the seams it already had
2,409 lines and 48 methods in one class, the largest file in the repository.
The seams were not hard to find: the class was already assembled from mixins
(`SurfaceConfigurationMixin`, `SurfaceProgressMixin`), each in its own module
with a docstring saying why those methods belong together. This follows that
pattern for the eight groups still inside.

Outbound delivery, target resolution, native interactions, routing, the
conversation link lifecycle, credentials, ingress preparation and inbound
message persistence. What is left in `ingress_service` is the class, its
constructor, and the three entry points a caller actually names.

`ResolvedSurfaceRoute` and `SurfaceEgressTarget` move to a leaf module. They
cannot stay where they were: every mixin that needs one is imported *by* that
module, so reaching back for the type would be a cycle. The egress target loses
its leading underscore in the move -- a private name half the package imports is
private only by spelling.

Pure relocation. Three tests needed retargeting because they reach for a symbol
where it used to live: two monkeypatch a call site that moved, and one imported
`_ask_user_request_dict` through `ingress_service` when
`pending_interaction_resume` is where it is defined. Nothing else changed, and
5,328 unit tests pass either side of the move.

Complexity is untouched on purpose. The counts are per module, so relocating a
method cannot move them -- that needs the functions decomposed, which is its own
commit rather than one hidden inside a move this size.
