+++ lemma-backend/app/modules/agent_surfaces/services/surface_inbound_message.py
@@ -1,0 +1,220 @@
+"""Recording what arrived, and folding in what arrived with it.
+
+The write half of ingress: persist the inbound message, and enrich it from
+things that are not the message itself -- a voice note that has to be
+transcribed first, recent channel history a group mention needs for context.
+"""
+
+from __future__ import annotations
+
+import asyncio
+from typing import Any
+
+
+from app.core.authorization.current import reset_current_context, set_current_context
+from app.core.authorization.factory import create_authorization_data_service
+
+from app.composition.surface_agent import ConversationService
+from app.modules.agent_surfaces.domain.ingress_context import (
+    SurfaceChatContext,
+)
+from app.modules.agent_surfaces.domain.ports import (
+    SurfacePlatformAdapterPort,
+)
+from app.modules.agent_surfaces.services.pending_interaction_resume import (
+    # Re-exported: ``_ask_user_request_dict`` still has a caller here (the
+    # native-interaction path) and a unit test that imports it from this module.
+    maybe_resume_pending_interaction,
+)
+from app.modules.agent_surfaces.services.surface_file_ingest_service import (
+    IngestedAttachment,
+)
+from app.core.log.log import get_logger
+
+logger = get_logger(__name__)
+
+_CONVERSATION_TITLE_MAX_LENGTH = 120
+# Recent thread/channel messages fetched per run for group-mention continuity.
+_CHANNEL_CONTEXT_LIMIT = 15
+
+
+class SurfaceInboundMessageMixin:
+    async def _commit_inbound_message(
+        self,
+        context: SurfaceChatContext,
+        message_text: str,
+        metadata: dict[str, Any],
+    ):
+        """Persist the inbound message / resume the paused run in a short UoW."""
+        if self._uow_factory is not None:
+            if self._conversation_service_factory is None:
+                raise RuntimeError("Conversation service factory is unavailable")
+            async with self._uow_factory() as uow:
+                conversation_service = self._conversation_service_factory(uow)
+                return await self._write_inbound_message(
+                    context, message_text, metadata, uow, conversation_service
+                )
+        else:
+            if self.uow is None or self.conversation_service is None:
+                raise RuntimeError("Conversation service is unavailable")
+            return await self._write_inbound_message(
+                context, message_text, metadata, self.uow, self.conversation_service
+            )
+
+    async def _write_inbound_message(
+        self,
+        context: SurfaceChatContext,
+        message_text: str,
+        metadata: dict[str, Any],
+        uow,
+        conversation_service: ConversationService,
+    ):
+        if context.pod_id is None:
+            raise ValueError("Surface chat context requires a pod")
+        # An empty inbound is never something a person sent — it means a body we
+        # failed to fetch or parse. Starting a run on it burns a model call and
+        # produces an answer to nothing, which reads to the sender as the agent
+        # ignoring them. Every inbound Resend email looked like this.
+        if not str(message_text or "").strip():
+            logger.warning(
+                "agent_surfaces.ingress_service.inbound_message_empty.degraded",
+                conversation_id=str(context.conversation_id),
+                platform=context.platform,
+            )
+            return None
+        auth_ctx = await create_authorization_data_service(uow).build_user_context(
+            user_id=context.user_id,
+            pod_id=context.pod_id,
+        )
+        token = set_current_context(auth_ctx)
+        try:
+            # If the run is paused on an ask_user, treat this inbound text as the
+            # answer and resume — rather than starting a new message/run. This is
+            # how the formatted-text fallback (and any "type your own" reply) gets
+            # back into the run as a structured answer.
+            if not await maybe_resume_pending_interaction(
+                context, message_text, conversation_service=conversation_service
+            ):
+                return await conversation_service.add_user_message_and_start_run(
+                    conversation_id=context.conversation_id,
+                    user_id=context.user_id,
+                    content=message_text,
+                    pod_id=context.pod_id,
+                    agent_name=context.agent_name,
+                    message_metadata=metadata,
+                )
+            return None
+        finally:
+            reset_current_context(token)
+
+    async def _fetch_channel_context(
+        self,
+        *,
+        adapter: SurfacePlatformAdapterPort,
+        context: SurfaceChatContext,
+        credentials: dict[str, Any],
+    ) -> list[dict[str, Any]]:
+        """Best-effort recent thread/channel messages for a group mention, as a
+        list of ``{author, text, ts}`` dicts. Fetched fresh per run; never raises."""
+        try:
+            messages = await adapter.fetch_thread_context(
+                credentials=credentials,
+                event=context.event,
+                limit=_CHANNEL_CONTEXT_LIMIT,
+            )
+        except Exception:
+            logger.debug(
+                "agent_surfaces.ingress_service.surface_channel_context_fetch_platform.diagnostic",
+                conversation_id=context.conversation_id,
+            )
+            return []
+        return [m.model_dump(mode="json") for m in messages][:_CHANNEL_CONTEXT_LIMIT]
+
+    async def _transcribe_voice_attachments(
+        self,
+        *,
+        ingested: list[IngestedAttachment],
+        original_text: str | None,
+        metadata: dict[str, Any],
+    ) -> str:
+        """Transcribe inbound voice notes and fold them into the message text.
+
+        The transcript becomes the user's words so the agent just reads text.
+        Join rules: caption + voice → both; voice-only → transcript alone;
+        several voices → labelled concatenation. A failed/oversize/empty voice
+        falls back to ``[voice message]`` (so a voice-only message is never an
+        empty prompt) while the saved audio file stays available. Provenance
+        (path + transcript + language) is recorded in ``metadata``.
+        """
+        original = (original_text or "").strip()
+        audio_present = [item for item in ingested if item.is_audio]
+        if not audio_present:
+            return original
+
+        to_transcribe = [item for item in audio_present if item.audio_bytes is not None]
+        provider = None
+        if to_transcribe:
+            try:
+                from app.composition.surface_agent import get_speech_provider
+
+                provider = get_speech_provider()
+            except Exception as exc:
+                # Voice notes arrive untranscribed from here on, which is a
+                # user-visible degradation — so it stays a warning, but it has
+                # to name the failure. The previous line carried no fields at
+                # all, so it could only report that something was wrong.
+                logger.warning(
+                    "agent_surfaces.ingress_service.speech_provider_unavailable",
+                    error_type=type(exc).__name__,
+                )
+                provider = None
+
+        async def _one(item: IngestedAttachment) -> tuple[IngestedAttachment, Any]:
+            try:
+                result = await provider.transcribe(
+                    item.audio_bytes, mime=item.mime or "audio/ogg"
+                )
+                return item, result
+            except Exception:
+                return item, None
+
+        results: list[tuple[IngestedAttachment, Any]] = []
+        if provider is not None and to_transcribe:
+            results = list(
+                await asyncio.gather(*[_one(item) for item in to_transcribe])
+            )
+
+        transcripts: list[str] = []
+        provenance: list[dict[str, Any]] = []
+        for item, result in results:
+            text = (getattr(result, "text", "") or "").strip()
+            if text:
+                transcripts.append(text)
+                provenance.append(
+                    {
+                        "path": item.path,
+                        "text": text,
+                        "detected_language": getattr(result, "detected_language", None),
+                        "duration_seconds": getattr(result, "duration_seconds", None),
+                    }
+                )
+            else:
+                provenance.append({"path": item.path, "text": "", "failed": True})
+        if provenance:
+            metadata["voice_transcripts"] = provenance
+        if not transcripts:
+            metadata["voice_transcription_failed"] = True
+
+        if not transcripts:
+            combined = "[voice message]"
+        elif len(transcripts) == 1:
+            combined = transcripts[0]
+        else:
+            combined = "\n\n".join(
+                f"[Voice {index}]\n{text}"
+                for index, text in enumerate(transcripts, start=1)
+            )
+
+        if original:
+            return f"{original}\n\n{combined}"
+        return combined
