+++ lemma-backend/architecture-baseline.json
@@ -99,7 +99,6 @@
     "app/modules/agent_surfaces/platforms/teams/service.py": 618,
     "app/modules/agent_surfaces/platforms/whatsapp/service.py": 623,
     "app/modules/agent_surfaces/services/event_receiver_service.py": 669,
-    "app/modules/agent_surfaces/services/ingress_service.py": 2409,
     "app/modules/agent_surfaces/services/surface_service.py": 954,
     "app/modules/apps/services/app_service.py": 628,
     "app/modules/connectors/services/connector_operation_service.py": 665,
